requirements.txt
fastapi==0.128.3
uvicorn==0.40.0
httpx==0.28.1
redis==7.1.0
google-genai==1.62.0
pillow==12.1.1
python-dotenv==1.2.1
